Use Id to represent identifiers, not String/primitives

+++ atlas-core/src/main/java/org/atlasapi/segment/Segment.java
@@ -17,7 +17,7 @@ public class Segment extends Described {
     private Publisher publisher;
 
     public SegmentRef toRef() {
-        return new SegmentRef(checkNotNull(this.getId().longValue(), "Can't create reference for segment without ID"));
+        return new SegmentRef(checkNotNull(this.getId(), "Can't create reference for segment without ID"));
     }
 
     public SegmentType getType() {
